Add NGINX vhosts configurations and extra phppgadmin
Based on the suggested NGINX configurations by the CMS-dev
community I created the ones for this bulky box. I also added
php-fpm extra configurations to render phppgadmin UI.

By the way phppgadmin is not a required dependency but I'm leaving
it in the Vagrant box for development and debugging purposes.

PS: I started with the documentation, is just a draft at this
moment