Distribute environment in 3 VMs w/docker
Since Ubuntu Bionic Vagrant image comes with a 10GB hard disk
limit it was not able to support the load of the whole CMS+Orbits
ecosystem, so I had to distribute the containers on 3 independent
virtual machines. One for Orbits, other for CMS and the last one
for data storage services, on average those take about 15GB of HD
and about 4.25GB of RAM from the virtual machine host.